Minimum of the Maximum group stat size
The groups maximum group size reported in group stat summary lines now reports the minimum size of the maximum group stat size for the word selections that are identified as yielding the maximum number of groups. Sometimes it would not always be apparent that word selections differ in this regard.